windows install: drop DelphiViewer build
From now on, the users of the Delphi can use the Odemis Viewer, with
exactly the same functionalities. This actually has been done in
practice for a couple of years.

We might eventually reinstate it, or use similar code to generate a
different version of the Odemis Viewer, so we keep the code around.