util driver: Never try to access the back-end if the file is not present
We used to just rely on Pyro to quickly find out it's not possible,
however, this leaves an ERROR log. In practice this meant that almost
every start of the backend started with:
ERROR   core:401:       cannot connect: [Errno 2] No such file or directory

It's a little confusing for the user. Especially in case there is really
an error, it's easy to be mislead and think this is the problem (as it's
the first error of the log).