gui model: remove label tool from acquisition tabs
The SECOM and SPARC acquisition tabs had the new label tool.
However, it's typically not really useful during acquisition. In
addition, it's currently not possible to save them, and having them
in this tab could easily disapoint the user when opening the acquired
file and noticing no label was saved.