odemisd modelgen: Minor fixes related to children
In case a component uses both children and dependencies, because it
needs both, it shouldn't warn that it's a mix of old and new children
style. => Detect such cases.

Also, if a wrapper component depends on a child of another component,
the container information wasn't available, which would show an
incorrect warning, and also could let the wrapper be executed on a
separate container.