gui xrc: Use "default font" instead of hardcoding "Ubuntu"
Use the special keyword for using the default font, and drop all the
settings which correspond the default settings.

In practice, as Odemis runs typically on Ubuntu, that's equivalent.
Unless of course, it runs on Windows, like for the Viewer.

This should allow the interface to follow more the standard user
interface.
On Windows, it might even avoid some weird popup asking the user to pick
a font... but that still needs to be tested.