gui cont tabs: extra fix for StaticText clipped on GTK3
Commit b32abd3fc6c (fix big StaticTexts clipped on GTK3) worked... often.
It seems that sometimes, if the computer is a little slow, it's too
early. So now, call the function twice at different interval, so that if
it was too early, we still do it.

I've done quite some search on what (event) needs to happen first before this
can work, but couldn't find anything conclusively.
So here we are, with some extra code that helps make this bug workaround even uglier.