acq stream sync: improve time estimation updates
We had an extra time cost for the assembling of the data at the end.
With assembling the data live, this is not needed anymore.
Especially, on CL intensity streams, which can have 1000x1000 px, 1 ms
per pixel was very excessive. This caused the time left to always be
very pessimistic.