acq stream sync: fix SEMMDStream live update thread failing sometimes
It could happen at the end of an acquisition that the _live_data is
already reset, but the acq_done not yet set. This would cause an
IndexError, which stops the thread.

For the acquisition, it's not a big deal, but future acquisitions won't
have live update anymore.

=> Use current_scan_area as None to indicate that the acquisition is
over to the live update thread.

We also make sure to only update the image if the acquisition is not
over, like for the standard live update thread. Otherwise, we can end-up
with an image set, but no raw, which confuses .getBoundingBox().