plugin tileacq: on SPARC, acquire all the streams
It would automatically stop as soon as it finds one acquisition stream.
If there are multiple streams, it should just acquire *all* of them.

Also show the last tile acquire at the end of the acquisition, instead
of not showing anything, if no stitching is done.
Also show a popup message, to confirm everything went fine, instead of
just closing the window.