driver ueye: don't worry if the shutter mode cannot be changed
On some camera (or version of libueye ?), the device capabilities
indicate that it's possible to change the shutter mode, and whenever
reading or writing it, it fails with a "NOT SUPPORTED" error.

=> Don't worry about this, and just conclude that finally it's not
supported.

Also log the hardware version at startup (handy for debugging).

Also fix the test cases which only worked when there was no hardware!