LinkedHeight Actuator wrapper
The following is an implmentation of LinkedHeightActuator and its child LinkedHeightFocus. The goal of this wrapper is to connect the lens Z and stage/sample. It ensures that the top of the lens doesn’t collide with the bottom of the stage/sample. It also allows to handle the main need that moving the stage Z should also move the focus Z.