model vattribute ListVA: force the value as a NotifyingList even if the setter changes the type
We used to first convert the value into a NotifyingList, and then pass
it to the setter. Normally the setter should be careful to keep it as
is. However, it's a pain to write in the setter. So, the nicer way is
convert the value to a NotifyingList after the setter is called.

The drawback is that the setter cannot anymore rely on the input to be
a NotifyingList, but it's just an Iterable, in general. No setter should
have made the distinction anyway.