driver semcomedi: support scanning_ttl high_enabled option
When the TTL has to be inverted for the scanning signal that doesn't
mean it should also be inverted when forced to True/False. It entirely
depends on what the signal means. For example, for an "external" signal,
we want it during for scanning, but still high when set to True.

=> Add another optional value in scanning_ttl