gui log: ensure the log shows the content of the variable when logged
We were delaying the formatting of the string to when it would be shown,
which can be delayed by up to 0.2s. For complex variables (eg, list,
VA), this could mean that if the variable has been changed in-between.
This is quite confusing.
In such case, the log in the GUI wouldn't be the same as in the
odemis-gui.log file.

=> Format the string immediately when logged.