acq align spot: CenterSpot does not need to move precisely when using the stage
This avoids trying too hard to move the spot to the center, until
essentially it gives up because of too many steps.

Also fix a small issue when the shift (returned by the function)
corresponds to the previous shift, before the last move. (typically that
value would very small, so not a big change)

Also add support for storing the images using "logpath" parameter