driver simulated: simulated.Light shouldn't be binary
It should simulate a light which has a continuous control between min
and max power. This is actually used by some installations, to report
the power of some light which is not software controlled.

Also, if there is a need for a binary light, it's still possible to use
the light.Light with a simulated power supplier.