util transform: check the type of the init arguments
There was very little check on the arguments passed.
In particular, scale should be two values, excepted for the SimilarityTransform
where it's a single value. On the AffineTransform, passing a single
value kind-of work, but the inverse returns something wrong.

Also update the test cases to always pass double value when required.