config: switch to Python 3 by default
Change the shebang header to explicitly call Python 3.

And for cases where the PYTHON_INTERPRETER variable is used, default to
Python 3.

We still support running Odemis on Python 2, but this is now deprecated.

Also remove an old test script, which has been broken for a few years,
and has even less chances to work on Python 3.