driver pigcs: fix Python 3 issue for hardware with weird multi-line
The code to backtrack and reconsider a previous line as a begining of a
multi-line uses the already decoded text, so it's a string, not bytes.
On Python 3, this matters.

This was not detected in the test cases, with the simulator and
other hardwares, as it only happens for very few old version of the PI
controllers.