driver andorcam3: fix USB3 check sometimes not checking
If the computer has a USB device with a serial number which is weirdly
encoded, then the code would fail to decode it (as ascii). In this case,
instead of skipping the device (because it's not a Zyla), it would
completely stop looking for the Zyla, and so never check whether USB3 is
active.

=> just skip the device in such case.