model component: forbid overriding VAs on a Proxy
When writing Python code, due to the weird syntax of VAs, it's easy to
override a VA instead of actually setting a value. This should hopefully
catch a few bugs a bit earlier.

We could also do a similar behaviour on a Component, but there might be
some specific cases where this is acceptable. Maybe we could acccept
overriding a VA only with another VA... but for now I don't dare trying.