gui comp file: On showing filename for saving, adjust extension based on format
When using the file dialog to pick a new name, for saving a file, the
file extension is not automatically adjusted based on the file format
selected by the user.

Other places in Odemis automatically adjusted the extension based on
file format, but not the FileBrowser. Now, also adjust here, so that the
final file name fits the file format selected by the user.

This is not used anywhere in the GUI, but it is used in many plugins.
This made selecting a different file format in the plugins very
difficult, and many users ended up using TIFF unwillingly.