driver phenom: do not ask for the progress area selection at random time
It seems that sometimes, the SOAP messages get mixed up, and if an image
is being received, it will be passed to this caller. Maybe this is a
more general issue and we should have a lock to protect the calls to
SOAP... for now just make sure the acquisition thread is stopped before
sending this message.

It would fail with such error:
Exception: <image/> not mapped to message part