acq stream live: Improve light power control
While setting up the excitation and emission, no image is being
acquired. As the emission actually exposes the sample, it's best to
set it up last, so that if the excitation setup is long, it's not
exposed.

Also ensure that setting the emission power is done in one step, instead
of first turning off everything and then turning on a channel.
It shouldn't have much effect, just a tiny bit faster.

Also fix the brightfield power setup to not always set the power to the
maximum when starting to play the stream. Also adjust all the channels
when changing the power, not just the first one.
Note that this is more a theoretical issue, as currently all the
hardware for brightfield only have 1 channel.