driver andorshrk: option to workaround slit reference issue
One spectrograph's slit had the reference switch not working.
As we don't need a very precise slit opening (because the user can
visually check), we can work around it by forcing the slit to go to the
maximum opening before every move. This has a similar effect as the
referencing as it allows to know where is the slit.