driver zeiss: support for scan rotation
Mostly handy to check if the scan rotation of the system has been
changed. Also force the rotation active when the angle > 0.
There doesn't seem to be a way to read if the rotation is active, so we
just have to assume it's active.
That means that in some corner cases it might cause so higher-level code
to activate the rotation if it does:
ebeam.rotation.value = ebeam.rotation.value

Maybe that's what the user wants... maybe not.

Also fix log of the focus component which said "stage".

Also adjust the simulator to return "E" values when very small values
are set, as the real hardware.