gui SPARC: leave the bottom legend enabled during acquisition
With the new live view during acquisition, it's important to be able to
adjust the SEM/CL merge ratio. Ideally, we'd leave the entire viewport
enabled, with only restriction on the movement and zoom, but for now,
let's keep it simple, and just re-enable the bottom legend after
disabling the whole viewport.