odemisd: wait until component is fully stop before stopping its dependencies
The component was remove from the dependency graph just before call
.terminate() .
That meant that as soon as any other component has finished, it's also
concidered stopped. On large models, that meant that pretty much
everything would stop simultaneously, which was not the goal.
Especially, this would cause issues when the power supplier is stopped
too early.

=> Wait until .terminate() is finished to remove the component from the
dependency graph.