driver: add call to super in .terminate()
The Component class actually does some useful things on .terminate(),
but it can only do it if we call it. Admitedly, typically it all happens
at the backend termination, so it's not a big loss, but let's do it
properly.

Note that there are other drivers that could have it, but I'm just doing
it for the most used drivers for now.