add division from the future
Now that we are mostly always using Python 3, it's not so useful, and
most of the code concerned probably didn't do any division. However, for
the sake of it, and because we still rely on Python 2 for fallback,
we should still follow the old rule to always use the divisions from the future.