gui util img: Fix insert_tile_to_image() when image is totally out of range
It was supposed to be handled, but in some cases, it wasn't, and we
would get errors like:
Traceback (most recent call last):
  File "/home/meteor/development/odemis/src/odemis/util/__init__.py", line 453, in _li_thread
    f(*args, **kwargs)
  File "/home/meteor/development/odemis/src/odemis/gui/cont/views.py", line 445, in _onNewImage
    self._update_ovv()
  File "/home/meteor/development/odemis/src/odemis/gui/cont/views.py", line 460, in _update_ovv
    self.im_opt = insert_tile_to_image(img, self.im_opt)
  File "/home/meteor/development/odemis/src/odemis/gui/util/img.py", line 2533, in insert_tile_to_image
    img.rescale_hq(tile, (tile_sz_px[1], tile_sz_px[0], c))[diff_y:y_right, diff_x:x_right]
ValueError: could not broadcast input array from shape (0,3,3) into shape (22,3,3)

Instead of adjusting the code further, just switch to Cairo, like the
rest of the functions. Cairo handles everything out-of-the-box.
It even brings anti-aliasing, and support for blend mode if we want to
add it.