gui xrc: remove explicit font encoding
On Linux, UTF-8 is the default encoding so it's fine.
On Windows, it's not available, and that causes wxPython to show a weird
font selection window to the user, with the message:
"xrced: unknown encoding
No font for displaying text in encoding 'Unicode 8 bit (UTF-8)' found.
Would you like to select a font to be used for this encoding (otherwise the
text in this encoding will not be shown correctly) ?"

We don't need to specify the encoding, so let's just not do that.