driver smaract: updated cancellation of all running moves and state change on failed reference on init
If referencing on init, run the referencing in background (instead of blocking the starting of the backend), and in case of error, update the component state.
When terminating, make sure the stage stops moving immediately. Especially useful if the user cancels the backend start, and the system is still referencing.