gui tab SECOMAlign: Use and update lens aligner active position
The goal is to ensure that, on the CryoSECOM, the optical lens is positioned as close as possible from the last known aligned position after a referencing or after it has been parked away.

Move to the FAV_POS_ACTIVE position whenever the alignment tab is selected. In the SECOM alignment tab, whenever the lens aligner is moved, the FAV_POS_ACTIVE metadata is updated.