gui util img: fix merge_screen() only returning original im
In commit e7789884b3 (add format_bgra_to_rgb() function), merge_screen()
was updated to make use of the new function, but that was incorrectly
done, as the original code was a little confusing.

It needs to return the update *background*, not the input im.

Also update the docstring and test cases to indicate now it returns a
RGBA (instead of RGB or RGBA). The actual calling code in Odemis doesn't
care.