gui comp: Also show CLStreams using "screen" operator
They are also typically shown using tint, several of them together (eg,
using some RGB filters), and so it makes more sense to add the pixel
values, like for the FluoStreams, instead of averaging.

Note that in case there is only one CLStream, it's always shown first,
so it doesn't affect the behaviour in this case.