gui comp: allow changing merge ratio between two Fluo or CL streams
For simplification, we had said that such image is always shown
all together (ie, merge ratio == 0.5). It might still be handy in some
cases to be able to change the display ratio finely between two images.

Note that if the ratio != 0.5 and the background is not black, the
background will be seen a little bit. As this only happens in some
internal debug mode, that shouldn't matter.