gui conf data: add Seismic colormap to combobox
We already use it for some AR polarized image, so it's nice for the user
to be able to select it also for the other images.
It's mostly useful for "symetric" histogram, where there is some kind
of middle value, and the further away a value is, the more highlighted
it should be. In many cases, that's not useful for microscope data, but
maybe some users will find a use.