gui comp miccanvas: fix OverviewCanvas thumbnail not often updating
We cannot directly get the data from the _bmp_buffer. Although it's
supposed to be the memory buffer for the DC, it's not immediately
updated. As a matter of fact, on Linux, it's updated only when
switched out with another memory.

So, just use the normal ._get_img_from_buffer() to get the data.