driver simcam: ensure that the generated image is always within bounds
The new code to automatically move based on the stage position
introduced a bug which would make the data looked-up out of bound.

This was actually triggered during one of the test cases, but it would
just block, instead of failing, so we hadn't noticed it before.

=> Use similar code as the simsem, which now seems to be quite stable.
=> Improve the test case. Instead of just checking the internal method,
really create a new component, and check that the images are properly
generated.