driver simcam: remove blur_factor argument
This was used by a single simulator, to make the blur "stronger" when
moving the focus.
In addition, every DigitalCamera has a .depthOfField VA, which is automatically
computed based on the LENS metadata. This VA has a similar meaning. So
it makes sense that the simulated camera follows this VA to adjust the
blur factor.

=> Use depthOfField, and update the simulator with metadata which makes
the blur equivalent.