plugins ar_spectral: fix spectrum being inverted on some systems
We use the raw CCD, although normally it's the spectrometer, which wraps
the CCD that has an axis invertion if the wavelengths are the opposite
direction. Due to difference between hardware and possible extra mirror,
there is no convention of wavelength on the CCD. In practice this means
that there is 50% chance the wavelengths are inverted.

=> Automatically look at the spectrometer, and invert the wavelengths if
it does that.

Also remove some incorrect metadata which caused the data to be
sometimes not properly be displayed on the spacial image.