gui: small refactoring of OverviewAcquisitionDialog
The controller now returns the DataArrays instead of deciding itself what to do
with them.

Also clean-up a little bit the code by using a dedicated GUIData instead
of copying the one from the original tab.
Remove the use of a VA for the filename, as there is no user
interaction.
Remove the support of ROA, which is some left over from the acquisition
window.

Also, when starting acquisition change the FoV to fit the entire area.