gui util widgets ProgressiveFutureConnector: simplify range computation
We used to update the range and the value, as total/completed. It makes
sense, but it's actually not so simple, because if for some reason the
range is smaller than the previous value, GTK would complain.
So instead, keep things simple, and just report a ratio. Not between
0 and 1, as it needs an integer, but between 0 and 4096, which should
be more than enough resolution (that's more pixels that a really large
screen).