driver test actuator: Fix TestLinkedAxesActuator.test_stop() false positive
Due to the absolute tolerance being too large (100µm == 1 FoV!),
and the step size being small (so movement is slow), it considered
that the stop didn't work.
We could wait longer, but it's better to be more precise, and faster for
the movement.

Also clean-up a tiny bit of constants not used anymore, and use proper
test.assert_pos_not_almost_equal().

Now all tests seem to pass again.