driver test actuator: wait a little after canceling a move to check the position
The test case failed because it *thought* that the position hasn't
changed. However, the driver does update the position, it's just that
the .position VA is updated *after* the .cancel() returns. That's what is causing the failure.

We could force cancel() to wait for the move function returning, but that's hard to do,
and probably would be annoying in some cases.

So let's just accept that it behaves as-is.
We can modify the test cases to wait a bit before reading the stage position, and then it works.