gui conf: disable interpolation from the configparser
By default there is a fancy feature in ConfigParser to replace %xxx by
something else. We don't use it, and it prevents storing anything
containing % without escaping it.

In practice, if the user would put a % in the filename, the config file
couldn't be saved.

=> Disable the interpolation.

Also switch to just the Python 3 names, as we don't support Python 2
anymore. The SafeConfigParser will not work on newer versions of Python 3.