acq stream: reduce MDStream live update to 0.5 Hz
Doing the live update has some cost. In particular, for the MDSEMStream,
it means we break down the acquisition in smaller regions.
Starting/stopping the acquisition takes a bit of time, and also might
cause some hardware to go to "idle" mode, which takes extra time to
start again.

The live update is there for the user to see there is something
happening, and that the data is still good. Doing it every 2s, instead
of 3 times per second, should be more than sufficient.