driver zeiss: handle better delay in answer
Sometimes the SEM takes a bit of time to answer (eg, 3s). We would wait
at most 1s (and weirdly, always wait for an extra second).
This could cause incorrectly errors.

In addition, if this ever happened, the answer which would eventually
come wouldn't be flushed, causing the responses to be read after the
wrong command.
=> always try to read a whole line, and on error, flush the input.

Also change the simulator to use a separate thread, in order to be able
to simulate such error, and therefore check that the code works.