ASM API wrapper: extend VAs to check dependencies
The mbscanner.resolution VA, the mppc.cellTranslation VA and the mppc.cellCompleteResolution VA depend on each other.
So far only when starting to acquire a megafield and by then uploading the settings in the respective VAs, an error was raised if certain dependendies were not met.
Modified setters to check directly if dependencies are met or not and don't wait until acquisition is started.

rule: cell complete size >= cell effective size + max cell translation value