ASM API wrapper: modify filename VA to raise error when invalid name is provided
If an invalid name is provided raise an error which can be captured e.g. by the GUI. A warning in the backend, is not immediantly visible to the user. If the value is just overwritten, we end up overwritting an already acquired image. Raise an error, so the GUI e.g. can request a different name.
Remove ILLEGAL in name for characters as those are the actually allowed characters.