odemisd: automatically create the settings file if not present
In case /etc/odemis-settings.yaml didn't exists, it would fail to open
it and not store persistent data.

Because we were relying on a special code for the default value of
--settings, the "standard" code which happened when specifying
explicitly the path did work (mostly). So the test case didn't detect
it, because the test case used a special path.

=> Use the same code all the time.
=> When the file doesn't exist, explicitly create it. It can still fail
for other reasons (eg, not enough permissions, or directory doesn't
exists), in which case, it will not save the persistent data as before.