plugins ar_spectral: take into account readout rate for time estimation
The other streams already do it, but in this plugin we were taking shortcuts
and assumed that the CCD acquisition == exposure time.
However, with slow readout rates, and small exposure times, the readout
can take many time the exposure time. This caused the acquisition to
fail, because it timed out too early.

=> Use the same formula as the other streams.