driver technolution: check DataFlow.next() argument
Passing wronge field_num might be detected only when calling the
API, in a separate thread. So it's nice to immediately notice that
field_num is not going to work. That was actually the case with
passing a numpy array.

Also improve the test cases to simulate better what normal code would
do: it'd wait before each next.

Also remove a test case testing impossible double megafield acquisition.
It actually should be illegal, but as the ASM doesn't check and the
wrapper doesn't check we can pass several times the same field number.
Let's leave it allowed, but at least let's not check that it works, as
it shouldn't.