acq stitching tiledacq: add registrar and weaver arguments
These argument allow to adjust the exact behaviour of the tiling.
In particular, in some case we want to tile without overlap (because the
stage precisition is good enough). In such case, the default registrar
would get confused/sad. It'd fallback to IDENTITY, but generate a lot of
warnings on the way.

Also handle better the stage speed. Use the same way to estimate before
and during. In addition, give a lot more margin to move the stage. There
is no need to fail if the stage is 1s late (eg, the smaract 5DoF adds
typically a 1s to avoid vibration). All we care about it to not block
forever if the stage stops responding.