driver smaract 5DoF: add settle_time option
The shape of the stage is such that even when the actuators are at
final target position, the sample is still vibrating. So we need to wait
a little extra time. Allow the user to specify this, by roughly adding a
extra settle time.

To save time when the user is moving the stage around, we skip this
extra time if there is already another move requested.