driver semcomedi: fix non-square pixel size on init
When the max_res wasn't square, the pixel size would be initialised
incorrectly to non-square value. On the first magnification update,
it would be set properly. As in almost every case the user has to change
the magnification, this is rarely noticeable. Still, it's good that
the pixelSize is always correct.