cli: set-attr convert list to tuples
When reading a series of values, as we interpret it as YAML, it's always
considered a list. The TupleVAs do handle both. However, in some cases, like an enumerated VA,
it requires to get exactly a value in choices. A list will not match a
tuple. To handle such case, we detect the VA contains tuple, and convert
the list to tuple.