driver xt_client: change .resolution and .scale to VAEnumerated
Only some very specific resolutions are accepted by the XT server.
They are always full frame. So try to pass this information as well
as possible by using VAEnumerated, and setting resolution to read-only.

So changing the scan resolution (which actually only means changing the
pixel pitch), can now be only done by changing the scale, and it's
possible to know which one is allowed.

Note that this is not really the standard interface for .resolution and
.scale, so the GUI doesn't adjust well automatically. But acquisition
will be used mostly only on new code anyway.