driver xt_client test: extra option to run with the adapter simulator
Use the same was as the orsay test case: add an extra state to
TEST_NOHW=sim. This indicate we are testing agains the simulator.

In such case, all test cases are safe to run, so run them all.
Also fix a few test cases to be compatible with the simulator.