driver andorcam2: Add support for using SW trigger
Some cameras may support software trigger. In this case, it can save
around 20ms when starting the image. For short exposure times (eg, 20
ms), that's quite an improvement in frame rate.

One of the difficulty is that apparently on some cameras, not all
settings (cropping?) are compatible with the software trigger. So we
need to have some extra handling for this. Haven't seen such case yet
though...

So in total we have 3 modes now: non-synchronized, synchronized via
start/stop, and synchronized via software trigger.

To avoid code explosion, we first refactored to use a FSM style
(inspired by the avantes driver). It also supports switching between
synchronized/non-synchronized on the fly.

The garbage collector can take up to 20ms to run. It's not worthy to run
every time. On live mode, it doesn't have too much effect, as the next image would
be under acquistion during that time.
However, with the software trigger, if the next event came early, it would
cause a delay before starting the next frame.
=> Only do it every 10 times (like the andorcam3), while waiting for the
hardware.

Also extend the simulator to support software trigger mode.