driver-lakeshore: avoid false positive syntax error in runtest script
The script looks for commas in the error message. Usually this is a sign of a syntax error since arguments need to be passed by '%' instead of ','. However, in this case, the comma has a different meaning, so it's not an error. To avoid the false positive warning, just move the string to a variable.