driver: fix Exception non-unpicklable
We had managed to make sure all the exception were pickable, which is
necessary for the backend to transfer them to the client (via Pyro4).
However, it turns out that an object can be pickable, but not
unpickable... and we had not cheked for it.
This is because the unpickling mechanism is essentially a new
initialisation. So the arguments passed to it must be the same as the
one passed to the super (which stores them for picking).

In addition, with Futures, when an exception is raised, it's sent
back to the client as "one-way" call. So the backend wouldn't know
that unpickling the exception failed. That mained that not only the call
failed, but the client thought it was taking a very long time... and
block.

=> Fix all the exceptions which had this problem.
=> Add test cases to detect such errors.