acq stream ARRawProjection: projectAsRaw() handles small ebeam position changes as the rest
We have a special way to handle tiny differences (due to floating point
error) in the ebeam positions between different polarizations.

However, the projectAsRaw() function didn't care about it. This caused
files with such tiny differences to display fine, export fine as images,
but fail to export as CSV.

=> Use the same indirect access to the raw data via ._pos[] as the rest
of the functions.