dataio/acq csv: fix extra new line on Windows
This is a known behaviour of the csv module of Python: the file gets
\r\r\n lines endings if not opening it with "newline=''". My
understanding is that it's due to the standard file wrapper to add a
newline, and the CSV writer also adding a newline.

See: https://docs.python.org/3.6/library/csv.html#id3

Make it compatible with Python 2 too. It might still have a bug in
Python 2, but let's not try to fix it there. The Windows viewer only
runs on Python 3 anyway.