technolution driver: remove waiting time for fix, where we scan the first field image twice
We scan the first field image twice due to HW restrictions. Remove the waiting time between the first and second scan as it is not needed. The scan_field call is blocking, meaning it waits until the scan is finished. Thus, we do not need to wait for the scan.