gui cont acquisition: fix changing the screen brightness on Python 3
The output of a command is bytes, but the call to a command must be
string.
So if we take read the output of a command and pass it to another
command, we need to convert it.

Also fix the detection of screen names which contained a "-".

Also add small test cases.