driver smaract: Fix 5DOF stage rare timeout issue
A timeout error can happen if the movement of some axes pulled others slightly out of their position, and because they are not expected to move by the kinematics (e.g. the x-positioners in  case of a pure y movement), they go back to their position very slowly, causing the movement event to time out. As a fix for this, the same movement will be tried again.