driver piezomotor: fix timeout on inverted axes moves
As we read .position directly, but the internal position was already
inverted (for inverted axes), the values didn't match. When computing
the distance between current and target positions, if the move happened
to be near its opposite position (eg, moving from 0.01 to -0.01), it
would seem that the distance is very short, so the duration short too.
The actual move could also be moved incorrectly, in relative move only.

=> Read the position and apply the inversion to compare the same type of
values.

Also simplify the code by passing the shifts, which is what we really
care.
Also support computing the estimated duration on open-loop.
Also make sure to stop all the axes from moving as soon as possible in
case of cancellation (instead of stopping just one).