gui comp PixelValueOverlay: show value under cursor immediately after activating it
When activating the overlay with Ctrl+U, it would display the value only
after moving the mouse a bit. That's confusing, as it feels it didn't
get activated.

=> immediately read the mouse position, so that we can display the
value.

Also use cnvs.Refresh() instead of cnvs.request_drawing_update(),
as it avoid redrawing the whole buffer, while all we need is a repaint.