gui comp viewport/miccanvas: Fold view.has_stage() into CAN_MOVE_STAGE
Now that we have the CAN_MOVE_STAGE ability, we can use it also
where we used to check if the view has a stage.

At init, if the view doesn't have a stage, drop the CAN_MOVE_STAGE.
Everywhere else, use CAN_MOVE_STAGE.

This is more than cosmetic: now if we have explicitly dropped
CAN_MOVE_STAGE, even if there is a stage, the canvas behaves
really as if there is no stage. That's important in fit_to_content()
and also allowing to drag when paused.