driver-simcam: increase range of exposure time
The image "andorcam2-fake-spots.h5" which can be used for the simulator has an integration time of 19.6 s. This causes an error to be raised since the range only accepts values up to 10. Since this is a simulator, it shouldn't be a problem to increase the range to 20s even though such long exposure times are maybe not common in practice.