remove use of subdirectory VA and move setting the filename VA to run method
The subdirectory VA was removed from the technolution wrapper. The full path and filename are now just stored in the filename VA.
Fix bug: Only set the filename VA in the run function. Multiple acquisition tasks can be generated while not being executed yet. Then the filename VA was overwritten though the first ROA was not yet acquired and would be stored under the id of the next ROA. To prevent this move, it to the run function, as the executor only runs one task at a time.