ENZEL: Add support for Features
Add a new tool TOOL_FEATURE in the localization tab, which allows to add/move features (bookmarks) around.

Features are shown as markers on the viewports. There is a toggle button to show/hide them. They have a few properties such as status, Z positions which can be edited directly. They also have a set of static streams associated to them.

All the Features information is stored in a dedicated JSON file in the project folder. It's automatically updated whenever the features are changed.