Add zstack acquisition manager for ENZEL
New acquisition manager which acquires over a set of Z positions. For now each stream is acquired serially. We could extend the method signature to support multiple intertwined channel acquisition by specifying a zlevels containing only a single list, but we don't need it for now.

Extend the ENZEL localization tab to use the Zstack acquisition when a checkbox is toggled in.
Also add a helper function to compute Z levels based on the current position and min/max/zstep.