driver simulated GenericComponent: improve VA class guessing
The original code was based on guessing the name of the class.
It tried to be too much creative. For instance, it would make a BooleanEnumerated, which do not exist.
VAEnumerated is always safe anyway.
For lists with range, it'd guess a ListContinuous, while only
TupleContinuous exists.

=> Explicitly code every type of VA supported. It's not that long.

Also force passing the default value.

Also handle choices with a set, instead of a dict.

Also fix error message.